+++ packages/components/src/actions/IconButton/index.tsx
@@ -10,7 +10,7 @@ import {
   getSharedButtonStyles,
 } from '../../primitives';
 import { useSharedPress } from '../../primitives/Button/useEvent';
-import { NATIVE_HIT_SLOP } from '../../utils';
+import { NATIVE_HIT_SLOP } from '../../utils/getFontSize';
 import { Tooltip } from '../Tooltip';
 
 import type { IButtonProps, IIconProps, IKeyOfIcons } from '../../primitives';
