+++ packages/components/src/content/NumberSizeableText/index.tsx
@@ -8,7 +8,7 @@ import type { INumberFormatProps } from '@onekeyhq/shared/src/utils/numberUtils'
 import { numberFormatAsRenderText } from '@onekeyhq/shared/src/utils/numberUtils';
 
 import { SizableText } from '../../primitives';
-import { getFontSize } from '../../utils';
+import { getFontSize } from '../../utils/getFontSize';
 
 import type { ISizableTextProps } from '../../primitives';
