+++ packages/components/src/forms/Checkbox/index.tsx
@@ -10,7 +10,7 @@ import platformEnv from '@onekeyhq/shared/src/platformEnv';
 import { Divider } from '../../content';
 import { ListView } from '../../layouts';
 import { Icon, Label, SizableText, XStack, YStack } from '../../primitives';
-import { NATIVE_HIT_SLOP } from '../../utils';
+import { NATIVE_HIT_SLOP } from '../../utils/getFontSize';
 
 import type { ILabelProps, IXStackProps, IYStackProps } from '../../primitives';
 import type { IFormFieldProps } from '../types';
