+++ packages/components/src/index.tsx
@@ -8,4 +8,3 @@ export * from './layouts';
 export * from './primitives';
 export * from './utils';
 export * from './types';
-export * from './utils/DebugRenderTracker';
