+++ packages/components/src/utils/sidebar.ts
@@ -1,0 +1,3 @@
+export const MAX_SIDEBAR_WIDTH = 208;
+
+export const MIN_SIDEBAR_WIDTH = 72;
